+++ NEWS.md
@@ -4,6 +4,8 @@
 - Update internal calls to `GetAssayData()` to use `layer` instead of `slot` (#160)
 - Update Matrix version to 1.6-2 (#164)
 - Change layer-saving in `SaveSeuratRds()` to move all layers instead of just those in `tempdir()` (#169)
+- Update internal calls to `SetAssayData()` to use `layer` instead of `slot` (#171)
+- Replace internal calls of `FilterObjects()` to `.FilterObjects()` (#171)
 
 # SeuratObject 5.0.0
 ## Added